# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/usuarios/{idUsuario}":{"get":{"tags":["usuario-controller"],"operationId":"buscarUsuarioPorId","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}}}}},"put":{"tags":["usuario-controller"],"operationId":"editarUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}}}}},"delete":{"tags":["usuario-controller"],"operationId":"deletarUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}}}}}},"/usuarios":{"get":{"tags":["usuario-controller"],"operationId":"listarUsuariosPaginado","parameters":[{"name":"pagina","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"tamanho","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUsuarioDto"}}}}}},"post":{"tags":["usuario-controller"],"operationId":"salvarUsuario","parameters":[{"name":"host","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDto"}}}}}}},"/digitos-unicos":{"post":{"tags":["calculo-digito-unico-controller"],"operationId":"obterDigitoUnico","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalculoDigitoUnicoDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CalculoDigitoUnicoDto"}}}}}}},"/digitos-unicos/usuario/{idUsuario}":{"get":{"tags":["calculo-digito-unico-controller"],"operationId":"buscarCalculosDigitoUnicoPorIdUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CalculoDigitoUnicoDto"}}}}}}}}},"components":{"schemas":{"CalculoDigitoUnicoDto":{"required":["inteiroParaExtracaoDigitoUnico"],"type":"object","properties":{"inteiroParaExtracaoDigitoUnico":{"type":"integer","format":"int32"},"numeroDeVezesParaRepeticaoDoInteiroParaExtracaoDigitoUnico":{"type":"integer","format":"int32"},"resultadoExtracaoDigitoUnico":{"type":"integer","format":"int32"},"usuarioCalculoDigitoUnico":{"$ref":"#/components/schemas/UsuarioDto"}}},"UsuarioDto":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int64"},"emailUsuario":{"type":"string"},"nomeUsuario":{"type":"string"},"calculosDigitoUnico":{"type":"array","items":{"$ref":"#/components/schemas/CalculoDigitoUnicoDto"}}}},"PageUsuarioDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/Sort"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/Pageable"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sort"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"}}},"Sort":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}